// src/app/shared/models/bank-account.model.ts
export class BankAccount {
  customer_bank_account_id: string;
  beneficiary_fname: string;
  beneficiary_lname: string;
  address: string;
  country_id: string;
  state: string;
  province: string;
  zip_code: string;
  bank_name: string;
  bank_address: string;
  bank_country: string;
  bank_state: string;
  bank_city: string;
  bank_branch: string;
  bank_province: string;
  bank_zip_code: string;
  bank_account_no: string;
  swift_code: string;
  title: string;
  status_name: string;
}