// src/app/shared/models/Ewallet.model.ts
export class ewallet{
    ownerName: string;
    ewalletType: string;
    walletAddress: string;
    billingEmail: string;
    billingAddress: string;
    billingCity: string;
    billingCountry: string;
    billingProvince: string;
    billingZip: string;
    billingTelCountryCode: string;
    billingTel: string
}